pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.2.0"
description = "Personal AI Employee - Silver Tier: Multi-watcher automation with planning, approvals, and scheduling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gmail = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
all = [
    "personal-ai-employee[gmail,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/watchers", "src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
